Refactor view state
Refactor view state to allow state operations directly on the state class instead of through the interface file handlers
