+++ src/state/view/state/index.ts
@@ -1,5 +1,28 @@
-import { DEFAULT_VIEW_TRANSFORM, DEVICE_PIXEL_RATIO } from "consts"
+import {
+    DEFAULT_VIEW_OFFSET_Y,
+    DEFAULT_VIEW_SCALE,
+    DEFAULT_VIEW_TRANSFORM,
+    DEVICE_PIXEL_RATIO,
+    MAX_PIXEL_SCALE,
+    TRANSFORM_PIXEL_SCALE_DEBOUNCE,
+    ZOOM_IN_WHEEL_SCALE,
+    ZOOM_OUT_WHEEL_SCALE,
+} from "consts"
+import { debounce } from "lodash"
+import { board } from "state/board"
 import { GlobalState, RenderTrigger } from "../../index.types"
+import {
+    applyBounds,
+    DetectionResult,
+    detectPageChange,
+    getCenterOfScreen,
+    getCenterX,
+    getCenterY,
+    getPageSize,
+    toNextPage,
+    toPreviousPage,
+    zoomTo,
+} from "../util"
 import {
     LayerConfig,
     ViewTransform,
@@ -19,6 +42,81 @@ export class View implements GlobalState<ViewState, ViewSubscribers> {
     subscribers: ViewSubscribers = {
         viewTransform: [],
         layerConfig: [],
+    }
+
+    updateViewTransform(newTransform: ViewTransform): void {
+        const detectionResult = detectPageChange(newTransform)
+
+        if (detectionResult === DetectionResult.Next) {
+            newTransform = toNextPage(newTransform)
+            board.incrementPageIndex()
+        } else if (detectionResult === DetectionResult.Previous) {
+            newTransform = toPreviousPage(newTransform)
+            board.decrementPageIndex()
+        }
+        newTransform = applyBounds(newTransform)
+
+        view.setTransformState(newTransform)
+        this.checkPixelScaleDebounced(newTransform)
+    }
+
+    checkPixelScale(newTransform: ViewTransform): void {
+        const pixelScale = Math.min(
+            DEVICE_PIXEL_RATIO * Math.ceil(newTransform.scale),
+            MAX_PIXEL_SCALE
+        )
+
+        if (pixelScale !== this.getLayerConfig().pixelScale) {
+            this.setLayerConfig({
+                pixelScale,
+            })
+        }
+    }
+
+    checkPixelScaleDebounced = debounce(
+        this.checkPixelScale,
+        TRANSFORM_PIXEL_SCALE_DEBOUNCE
+    )
+
+    zoomCenter(isZoomingIn: boolean): void {
+        const newTransform = zoomTo({
+            viewTransform: view.getViewTransform(),
+            zoomPoint: getCenterOfScreen(),
+            zoomScale: isZoomingIn ? ZOOM_IN_WHEEL_SCALE : ZOOM_OUT_WHEEL_SCALE,
+        })
+
+        this.updateViewTransform(newTransform)
+    }
+
+    resetView(): void {
+        this.updateViewTransform({
+            scale: DEFAULT_VIEW_SCALE,
+            xOffset: getCenterX(),
+            yOffset: DEFAULT_VIEW_OFFSET_Y,
+        })
+    }
+
+    private scaleWithFixedY(newScale: number, yFixed: number): number {
+        const { scale, yOffset } = this.getViewTransform()
+
+        return yOffset + yFixed / newScale - yFixed / scale
+    }
+
+    private rescaleAtCenter(newScale: number): void {
+        const newTransform = {
+            scale: newScale,
+            xOffset: getCenterX() / newScale,
+            yOffset: this.scaleWithFixedY(newScale, getCenterY()),
+        }
+        this.updateViewTransform(newTransform)
+    }
+
+    resetViewScale(): void {
+        this.rescaleAtCenter(1)
+    }
+
+    fitToPage(): void {
+        this.rescaleAtCenter(window.innerWidth / getPageSize().width)
     }
 
     setState(newState: ViewState) {
